requirements.txt
pydantic>=2.0.0
python-dateutil>=2.8.2
typing-extensions>=4.5.0
uuid>=1.30 
langchain
numpy 
sentence-transformers
pinecone
python-dotenv
boto3
fastapi
uvicorn
redis>=5.0.1
sse-starlette>=1.8.2